Fix buffers-tabline error when opening directories
This patch fixes an index out of range exception when opening vim with a
directory as a parameter. The issue is caused because there are no
buffers active when the plugin creates the tabline. Populating the
buffer list with one buffer solves the issue the buffer gets created
before the user gets shown the screen I assume.

The errors this fixes are the following, when calling lightline#tabline
with 'buffers' as a member and opening vim with `vim .`

```
E684: list index out of range: -1
E15: Invalid expression: [a:range[-1], [], [], [], '...', min([max([winwidth(0) / 40, 2]), 8])]
```